# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "嘉阅圈私域社群平台后端服务，提供内容社区、互动打卡、小组交流、私信通知、知识课程等核心能力。\n统一响应格式：{\"success\":bool,\"data\":any,\"message\":string,\"error\":{\"code\":string,\"message\":string}}",
        "title": "嘉阅圈后端 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3100",
    "basePath": "/api",
    "paths": {
        "/admin/scheduler/trigger": {
            "post": {
                "description": "需登录且需管理员。手动触发定时总结任务（daily/weekly/monthly），任务在独立协程中异步执行，接口立即返回，执行结果通过服务日志查看。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台-定时任务"
                ],
                "summary": "触发定时任务",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务类型（daily/weekly/monthly）",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "type 参数必须为 daily / weekly / monthly",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "403": {
                        "description": "仅平台管理员可触发定时任务",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "yjedu-reading-club-server_internal_pkg_response.ApiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "yjedu-reading-club-server_internal_pkg_response.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiError"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 \"Bearer \u003ctoken\u003e\"，token 通过 /api/auth/login/phone 或 /api/auth/weapp/login 获取。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}